+++ DangerFile
@@ -60,7 +60,7 @@ end
 
 #SWIFTLINT
 swiftlint.lint_all_files = true
-swiftlint.config_file = './.swiftlint.yml'
+swiftlint.config_file = '.swiftlint.yml'
 swiftlint.lint_files fail_on_error: true
 
 #TEST EVOLUTION CHECK:
