+++ Tests/Helpers/CodableTestObject.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct CodableTestObject: Codable, Equatable {
+    let name: String
+    let age: Int
+    let weight: Float
+}
